NodeJS workaround completed
