+++ src/cli/pdbe-images/scripting/browser.ts
@@ -30,9 +30,9 @@ async function initViewer() {
         collapseLeftPanel: false,
         pdbProvider: 'pdbe',
         emdbProvider: 'pdbe',
-        volumeStreamingServer:'https://www.ebi.ac.uk/pdbe/densities',
+        volumeStreamingServer: 'https://www.ebi.ac.uk/pdbe/densities',
         pixelScale: 1,
-        pickScale:  0.25,
+        pickScale: 0.25,
         pickPadding: 1,
         enableWboit: undefined,
         enableDpoit: undefined,
